requirements.txt
flask>=3.0.0
opencv-python-headless>=4.8.0
pillow>=10.0.0
requests>=2.31.0
pygame>=2.5.0
python-dotenv>=1.0.0
openai>=1.0.0
